requirements.txt
aiohttp~=3.8.6
Pillow~=10.0.1